# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pet Spa Booking API Documentation",
    "description": "RESTful API cho hệ thống đặt lịch Pet Spa",
    "version": "1.0.0",
    "contact": {
      "name": "Pet Spa Support",
      "email": "admin@petspa.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Pet Spa Booking API Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer token"
      }
    },
...
  },
  "paths": {
...
    "/admin/services": {
      "get": {
        "tags": ["Admin Services"],
        "summary": "Lấy danh sách dịch vụ (Admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "category_id", "in": "query", "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {"description": "Thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden - Admin only"}
        }
      },
      "post": {
        "tags": ["Admin Services"],
        "summary": "Thêm dịch vụ mới",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["serviceName", "categoryID", "price"],
                "properties": {
                  "serviceName": {"type": "string"},
                  "categoryID": {"type": "integer"},
                  "price": {"type": "number"},
                  "duration": {"type": "integer"},
                  "description": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Thêm thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden - Admin only"}
        }
      }
    },
    "/admin/services/{id}": {
      "get": {
        "tags": ["Admin Services"],
        "summary": "Lấy chi tiết dịch vụ",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Không tìm thấy"}
        }
      },
      "put": {
        "tags": ["Admin Services"],
        "summary": "Cập nhật dịch vụ",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Cập nhật thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Không tìm thấy"}
        }
      },
      "delete": {
        "tags": ["Admin Services"],
        "summary": "Xóa dịch vụ",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Xóa thành công"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Không tìm thấy"}
        }
      }
    },
...
  }
}